remerge ktmetrics(over)
