+++ pkg/ktexporter/fetchmetrics/fetchmetrics.go
@@ -1,20 +1,19 @@
-package fetch_metrics
+package fetchmetrics
 
 import (
 	"bytes"
 	"fmt"
-	"os"
 	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
 
-	jsoniter "github.com/json-iterator/go"
+	"os"
 
+	jsoniter "github.com/json-iterator/go"
 	"github.com/secretflow/kuscia/pkg/utils/nlog"
 )
 
-// ContainerConfig represents the config.json structure.
 type ContainerConfig struct {
 	Hostname string `json:"hostname"`
 }
@@ -103,6 +102,65 @@ func GetTaskIDToContainerID() (map[string]string, error) {
 	return taskIDToContainerID, nil
 }
 
+// ContainerStats holds the stats information for a container
+type ContainerStats struct {
+	CPUPercentage string
+	Memory        string
+	Disk          string
+	Inodes        string
+}
+
+// GetContainerStats fetches the container stats using crictl stats command
+func GetContainerStats() (map[string]ContainerStats, error) {
+	// Execute the crictl stats command
+	cmd := exec.Command("crictl", "stats")
+	var out bytes.Buffer
+	cmd.Stdout = &out
+	err := cmd.Run()
+	if err != nil {
+		nlog.Warn("failed to execute crictl stats", err)
+		return nil, err
+	}
+
+	// Parse the output
+	lines := strings.Split(out.String(), "\n")
+	if len(lines) < 2 {
+		nlog.Warn("unexpected output format from crictl stats")
+		return nil, err
+	}
+
+	// Create a map to hold the stats for each container
+	statsMap := make(map[string]ContainerStats)
+
+	// Skip the header line and iterate over the output lines
+	for _, line := range lines[1:] {
+		// Skip empty lines
+		if line == "" {
+			continue
+		}
+
+		// Split the line by whitespace
+		fields := strings.Fields(line)
+		if len(fields) < 5 {
+			nlog.Warnf("unexpected output format for line: %s", line)
+			return nil, err
+		}
+
+		// Extract the stats information
+		containerID := fields[0]
+		stats := ContainerStats{
+			CPUPercentage: fields[1],
+			Memory:        fields[2],
+			Disk:          fields[3],
+			Inodes:        fields[4],
+		}
+		// Add the stats to the map
+		statsMap[containerID] = stats
+	}
+
+	return statsMap, nil
+}
+
 // GetMemoryUsageStats retrieves memory usage statistics (virtual, physical, swap) for a given container
 func GetMaxMemoryUsageStats(pid string, cidPrefix string) (uint64, uint64, error) {
 	// Find the full CID based on the prefix
@@ -176,6 +234,68 @@ func getPhysicalMemoryUsage(cid string) (uint64, error) {
 
 	return physicalMemory, nil
 }
+
+func GetContainerNetIOFromProc(defaultIface, pid string) (recvBytes, xmitBytes uint64, err error) {
+	netDevPath := fmt.Sprintf("/proc/%s/net/dev", pid)
+	data, err := os.ReadFile(netDevPath)
+	if err != nil {
+		nlog.Warn("Fail to read the path", netDevPath)
+		return recvBytes, xmitBytes, err
+	}
+
+	lines := strings.Split(string(data), "\n")
+	if len(lines) < 3 {
+		nlog.Error("unexpected format in ", netDevPath)
+		return recvBytes, xmitBytes, err
+	}
+	recvByteStr := ""
+	xmitByteStr := ""
+	for _, line := range lines {
+		if line == "" {
+			continue
+		}
+
+		fields := strings.Fields(line)
+		if len(fields) < 10 {
+			continue
+		}
+
+		iface := strings.Trim(fields[0], ":")
+		if iface == defaultIface {
+			recvByteStr = fields[1]
+			xmitByteStr = fields[9]
+		}
+	}
+	if recvByteStr == "" {
+		recvByteStr = "0"
+	}
+	if xmitByteStr == "" {
+		xmitByteStr = "0"
+	}
+	recvBytes, err = strconv.ParseUint(recvByteStr, 10, 64)
+	if err != nil {
+		nlog.Error("Error converting string to uint64:", err)
+		return recvBytes, xmitBytes, err
+	}
+	xmitBytes, err = strconv.ParseUint(xmitByteStr, 10, 64)
+	if err != nil {
+		nlog.Error("Error converting string to uint64:", err)
+		return recvBytes, xmitBytes, err
+	}
+
+	return recvBytes, xmitBytes, nil
+}
+
+func GetContainerBandwidth(curRecvBytes, preRecvBytes, curXmitBytes, preXmitBytes uint64, timeWindow float64) (recvBandwidth, xmitBandwidth float64, err error) {
+	recvBytesDiff := float64(curRecvBytes) - float64(preRecvBytes)
+	xmitBytesDiff := float64(curXmitBytes) - float64(preXmitBytes)
+
+	recvBandwidth = (recvBytesDiff * 8) / timeWindow
+	xmitBandwidth = (xmitBytesDiff * 8) / timeWindow
+
+	return recvBandwidth, xmitBandwidth, nil
+}
+
 func findCIDByPrefix(prefix string) (string, error) {
 	cgroupDir := "/sys/fs/cgroup/cpu/k8s.io/"
 	files, err := os.ReadDir(cgroupDir)
